Implement Edit Product Page

+++ src/models/product.ts
@@ -78,18 +78,36 @@ export default class Product {
         });
     }
 
-    save(callback: (err: NodeJS.ErrnoException | null) => void) {
+    save(allowOverwrite: boolean, callback: (err: NodeJS.ErrnoException | null) => void) {
         Product.getProductJsonArrayFromFile((array) => {
-            array.push({
-                title: this.title, 
-                description: this.description,
-                cost: this.cost,
-                imageUrl: this.imageUrl,
-                id: this.id
-            });
-            fs.writeFile(Product.getFilePath(), JSON.stringify(array), 'utf8', (err) => {
-                callback(err);
-            });
+            if(allowOverwrite) {
+                const newArray = array.filter((p) => p.id !== this.id);
+                newArray.push({
+                    title: this.title, 
+                    description: this.description,
+                    cost: this.cost,
+                    imageUrl: this.imageUrl,
+                    id: this.id
+                });
+                fs.writeFile(Product.getFilePath(), JSON.stringify(newArray), 'utf8', (err) => {
+                    callback(err);
+                });
+            } else {
+                const search = array.filter((p) => p.id === this.id);
+                if(search.length > 0) {
+                    return callback(null);
+                }
+                array.push({
+                    title: this.title, 
+                    description: this.description,
+                    cost: this.cost,
+                    imageUrl: this.imageUrl,
+                    id: this.id
+                })
+                fs.writeFile(Product.getFilePath(), JSON.stringify(array), 'utf8', (err) => {
+                    callback(err);
+                });
+            }
         });
     }
